tckconvert: Store data for empty .vtk file
Due to truncation of any text following a hash character in a CI test line as introduced in #1274, a test for tckconvert introduced in #1268 no longer works, as a hash character is written to a temporary file via a printf call. This change instead adds the relevant file (a .vtk file wth no contents) to the test_data repository.