fixel2voxel: Update data to reflect #1352
Pull request #1352 involves altering the majority of outputs of the fixel2voxel command. Prior to this PR, these outputs are 4D images with two volumes, where the first volume contains the requested parameter and the second volume is empty; this is due to fixel2voxel using the fixel format index image as a template for the output image. This PR modifies these outputs to be 3D images, as one would expect instinctively based on the parameters being requested. This particular commit to the testing branch updates the relevant test data for this command, by using mrconvert to explicitly reduce the existing test data to the expected three dimensions.